requirement.txt
flask
python-dotenv
langchain
langchain-huggingface
langchain-community
faiss-cpu
langchain-google-genai
